+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog 
 
+## Version 1.4.0 - 2021-06-28
+
+- Added CreateDataset
+- Added ListDatasets
+- Added UpdateDataset
+- Added DeleteDataset
+- Added datasetId argument to CreateDocument
+- Added datasetId argument to ListDocuments
+- Added datasetId argument to DeleteDocuments
+
 ## Version 1.3.0 - 2021-05-21
 
 - Added CreateModel